refactor(crowdfund-ui): promote spike radial TreeView to canonical
Promotes the spike's D3 radial force-directed view (previously named
`RadialTreeView`) to be THE TreeView used by observer + committer apps.

Cleanup:
- Delete legacy xyflow-based TreeView.tsx (~760 lines).
- Rename RadialTreeView.tsx → TreeView.tsx (same content; git records
  as delete + rewrite because the new file content differs wholesale
  from the legacy file).
- Rename internal symbols: `RadialTreeView` → `TreeView`,
  `RadialTreeViewProps` → `TreeViewProps`.
- Drop the `RadialTreeView as TreeView` import alias in observer and
  committer App.tsx — they import `TreeView` directly now.
- Collapse duplicate exports in shared/index.ts to a single `TreeView`.
- Drop unused deps: `@xyflow/react`, `d3-hierarchy`, `d3-scale`
  (plus their `@types/*`). None of these have other callers in shared
  after the legacy removal. Lock file regenerated.
- Delete SPIKE_DEFERRED.md; its remaining content is superseded by
  a local parity-followups doc in .context/ (gitignored).

Parity with the legacy TreeView is not 100%: NodeDetail popover on
click, search-filter dim, "My wallet" zoom, and inviter-chain edge
emphasis are all missing and tracked as follow-ups. See the TreeView
file-header comment and the local parity-followups doc. These are
expected to land in subsequent commits on this branch before merge
into iskay/crowdfund-ui-polish.

Tests pass (143); typecheck clean across shared + observer + committer.
