+++ crowdfund-ui/packages/observer/src/App.tsx
@@ -14,7 +14,7 @@ import {
   StatsBar,
   TableView,
   SearchBar,
-  RadialTreeView as TreeView,
+  TreeView,
   AppShell,
   Button,
   Tabs,
