+++ crowdfund-ui/packages/shared/package.json
@@ -24,12 +24,9 @@
     "@types/d3-force": "^3.0.10",
     "@types/d3-selection": "^3.0.11",
     "@types/d3-zoom": "^3.0.8",
-    "@xyflow/react": "^12.10.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "d3-force": "^3.0.0",
-    "d3-hierarchy": "^3.1.2",
-    "d3-scale": "^4.0.2",
     "d3-selection": "^3.0.0",
     "d3-zoom": "^3.0.0",
     "ethers": "^6.15.0",
@@ -44,8 +41,6 @@
   "devDependencies": {
     "@hookform/resolvers": "^5.2.2",
     "@testing-library/react": "^16.3.2",
-    "@types/d3-hierarchy": "^3.1.7",
-    "@types/d3-scale": "^4.0.8",
     "@types/node": "^22.15.0",
     "fake-indexeddb": "^6.2.5",
     "react-hook-form": "^7.73.1",
