+++ crowdfund-ui/packages/shared/src/index.ts
@@ -107,9 +107,6 @@ export type { NodeDetailProps } from './components/NodeDetail.js'
 export { TreeView } from './components/TreeView.js'
 export type { TreeViewProps } from './components/TreeView.js'
 
-export { RadialTreeView } from './components/RadialTreeView.js'
-export type { RadialTreeViewProps } from './components/RadialTreeView.js'
-
 export type { RadialNode, RadialEdge, RadialGraph, AngleInfo } from './lib/radialLayout.js'
 export { buildRadialGraph, computeAngleMap } from './lib/radialLayout.js'
